requirements.txt
streamlit
google-genai
python-dotenv
Pillow
sentence-transformers
scikit-learn
numpy